+++ changelog.md
@@ -5,3 +5,4 @@
 - Installed [forever](https://www.npmjs.com/package/forever) to ensure continuously running of application.
 - Renamed the folder `helper` to `app`.
 - Renamed `app.js` to `server.js`.
+- Updated `getFileOptions` to dynamically know what fileOptions should be.